resume add to Nav component
