+++ src/component/Navbar.js
@@ -2,16 +2,21 @@ import React, { useState } from "react";
 import { FaBars, FaTimes } from 'react-icons/fa';
 import { Link } from "react-scroll";
 
+import resume_pdf from '../assets/resume/Ferny_Castro_Resume.pdf';
+
 function Navbar() {
     const [nav, setNav] = useState(false);
 
     const links = [
         {
             id: 0,
-            logo: [<span>
-                <a href=" ">
-                    <img src="https://see.fontimg.com/api/renderfont4/PKY87/eyJyIjoiZnMiLCJoIjo3NiwidyI6MjAwMCwiZnMiOjM4LCJmZ2MiOiIjRkZGRkZGIiwiYmdjIjoiIzM1M0Q0QiIsInQiOjF9/RkM/emotional-rescue-personal-use-regular.png" alt="Cursive fonts" /></a>
-            </span>],
+            logo: [
+                <span>
+                    <a href=" ">
+                        <img src="https://see.fontimg.com/api/renderfont4/PKY87/eyJyIjoiZnMiLCJoIjo3NiwidyI6MjAwMCwiZnMiOjM4LCJmZ2MiOiIjRkZGRkZGIiwiYmdjIjoiIzM1M0Q0QiIsInQiOjF9/RkM/emotional-rescue-personal-use-regular.png" alt="Cursive fonts" />
+                    </a>
+                </span>
+            ],
             target: 'navbar',
         },
         {
@@ -26,13 +31,25 @@ function Navbar() {
         },
         {
             id: 3,
+            link: 'links',
+            target: 'links',
+        },
+        {
+            id: 4,
             link: 'contact',
             target: 'contact',
         },
         {
-            id: 4,
-            link: 'links',
-            target: 'links',
+            id: 5,
+            link: 'resume',
+            target: 'resume',
+            // on click funtion to enable the download option for Resume
+            onClick: () => {
+                const link = document.createElement('a');
+                link.href = resume_pdf;
+                link.download = 'Ferny_Castro_Resume.pdf';
+                link.click();
+            }
         }
     ];
 
@@ -41,24 +58,32 @@ function Navbar() {
             <nav className="flex justify-between py-10 pb-8 items-center">
                 <span className="flex justify-center ">
                     <a href=" ">
-                        <img className="cursor-default" src="https://see.fontimg.com/api/renderfont4/PKY87/eyJyIjoiZnMiLCJoIjo3NiwidyI6MjAwMCwiZnMiOjM4LCJmZ2MiOiIjRkZGRkZGIiwiYmdjIjoiIzM1M0Q0QiIsInQiOjF9/RkM/emotional-rescue-personal-use-regular.png" alt="logo" /></a>
+                        <img className="cursor-default" src="https://see.fontimg.com/api/renderfont4/PKY87/eyJyIjoiZnMiLCJoIjo3NiwidyI6MjAwMCwiZnMiOjM4LCJmZ2MiOiIjRkZGRkZGIiwiYmdjIjoiIzM1M0Q0QiIsInQiOjF9/RkM/emotional-rescue-personal-use-regular.png" alt="logo" />
+                    </a>
                 </span>
 
                 <ul className="hidden md:flex items-center font-medium text-md cursor-pointer">
-                    {links.map(({ id, link, target }) => (
-                        <li key={id}
+                    {links.map(({ id, link, target, onClick }) => (
+                        <li
+                            key={id}
                             className="rounded-md ml-8 px-4 border-none text-gray-400 hover:underline hover:text-white hover:scale-105 duration-200"
                         >
-                            <Link
-                                to={target}
-                                spy={true}
-                                smooth={true}
-                                offset={-70}
-                                duration={500}
-                                onClick={() => setNav(false)}
-                            >
-                                {link}
-                            </Link>
+                            {onClick ? (
+                                <a href="/" onClick={onClick} download>
+                                    {link}
+                                </a>
+                            ) : (
+                                <Link
+                                    to={target}
+                                    spy={true}
+                                    smooth={true}
+                                    offset={-70}
+                                    duration={500}
+                                    onClick={() => setNav(false)}
+                                >
+                                    {link}
+                                </Link>
+                            )}
                         </li>
                     ))}
                 </ul>
@@ -66,8 +91,8 @@ function Navbar() {
                 <div onClick={() => setNav(!nav)} className="cursor-pointer pr-4 z-10 text-gray-400 md:hidden">
                     {nav ? <FaTimes size={25} /> : <FaBars size={25} />}
                 </div>
-
             </nav>
+
             {nav && (
                 <ul className="flex flex-col justify-center items-center absolute py-6 pb-10 top-0 left-0 w-full h-screen bg-gradient-to-b from-black to-gray-900">
                     {links.map(({ id, link, target, logo }) => (
@@ -89,8 +114,6 @@ function Navbar() {
                     ))}
                 </ul>
             )}
-
-
         </header>
     );
 }
